separated fragment and CircuitView add error method
