+++ app/src/main/java/com/study/thesuperiorstanislav/edaapp/main/MainFragment.kt
@@ -78,7 +78,7 @@ class MainFragment : Fragment(), MainContract.View {
     }
 
     override fun showData(circuit: Circuit) {
-        (circuitView as CircuitView).setCircuit(circuit)
+        circuitView.setCircuit(circuit)
         dialog?.dismiss()
     }
 
@@ -122,216 +122,6 @@ class MainFragment : Fragment(), MainContract.View {
         intent.addCategory(Intent.CATEGORY_OPENABLE)
         intent.type = "*/*"
         startActivityForResult(intent, READ_REQUEST_CODE)
-    }
-
-
-
-    class CircuitView : View {
-        constructor(context: Context) : super(context)
-        constructor(context: Context, attrs: AttributeSet) : super(context, attrs)
-
-
-        private val rect = Rect()
-        var renderHelper: RenderHelper? = null
-        private var circuit: Circuit? = null
-
-        var editEvent = EditEvent.DELETE_ELEMENT
-        var drawTouch = false
-        var startPoint = Point(-1,-1)
-
-        override fun onSizeChanged(w: Int, h: Int, oldw: Int, oldh: Int) {
-            super.onSizeChanged(w, h, oldw, oldh)
-            this.getLocalVisibleRect(rect)
-            renderHelper = RenderHelper(rect)
-        }
-
-        override fun onDraw(canvas: Canvas) {
-            super.onDraw(canvas)
-
-            renderHelper?.drawLines(canvas)
-
-            if (circuit != null && !renderHelper?.isMatrixInit!!)
-                renderHelper?.initDrawMatrix(circuit!!)
-
-            if (drawTouch)
-                renderHelper?.drawSelectedSquare(startPoint,canvas)
-
-
-            if (circuit != null)
-                renderHelper?.drawCircuit(circuit!!, canvas)
-
-
-        }
-
-        override fun onTouchEvent(event: MotionEvent): Boolean {
-            val x = event.x
-            val y = event.y
-
-            when (event.action) {
-                MotionEvent.ACTION_DOWN -> {
-                    when (editEvent) {
-                        EditEvent.VIEW -> TODO()
-                        EditEvent.ADD_ELEMENT -> TODO()
-                        EditEvent.ADD_NET -> TODO()
-                        EditEvent.ADD_CONNECTION -> TODO()
-                        EditEvent.MOVE_ELEMENT -> {
-                            if (drawTouch) {
-                                val endPoint = Point((x / renderHelper?.step!!).toInt(),
-                                        (y / renderHelper?.step!!).toInt())
-                                val obj = circuit?.listPins?.find {
-                                    it.getPoint() == startPoint }!!.getElement()
-                                if (renderHelper?.moveObject(obj, startPoint, endPoint)!!) {
-                                    obj.move(endPoint.x, endPoint.y)
-                                    drawTouch = false
-                                    invalidate()
-                                } else {
-                                    TODO()
-                                }
-                            } else {
-                                startPoint = Point((x / renderHelper?.step!!).toInt(),
-                                        (y / renderHelper?.step!!).toInt())
-                                if (renderHelper?.isTherePin(startPoint)!!) {
-                                    circuit?.listElements?.forEach { element ->
-                                        val point = element.getPins().find { it.getPoint() == startPoint }?.getElement()?.getPoint()
-                                        if (point != null) {
-                                            startPoint = point
-                                            drawTouch = true
-                                            invalidate()
-                                            performClick()
-                                            return true
-                                        }
-
-                                    }
-
-                                }
-                            }
-                        }
-                        EditEvent.MOVE_NET -> {
-                            if (drawTouch) {
-                                val endPoint = Point((x / renderHelper?.step!!).toInt(),
-                                        (y / renderHelper?.step!!).toInt())
-                                val obj = circuit?.listNets?.find {
-                                    it.getPoint() == startPoint
-                                }!!
-                                if (renderHelper?.moveObject(obj, startPoint, endPoint)!!) {
-                                    obj.move(endPoint.x, endPoint.y)
-                                    drawTouch = false
-                                    invalidate()
-                                } else {
-                                    TODO()
-                                }
-                            } else {
-                                startPoint = Point((x / renderHelper?.step!!).toInt(),
-                                        (y / renderHelper?.step!!).toInt())
-                                if (renderHelper?.isThereNet(startPoint)!!) {
-                                    drawTouch = true
-                                    invalidate()
-                                }
-                            }
-                        }
-                        EditEvent.DELETE_ELEMENT -> {
-                            startPoint = Point((x / renderHelper?.step!!).toInt(),
-                                    (y / renderHelper?.step!!).toInt())
-                            if (renderHelper?.isTherePin(startPoint)!!) {
-                                circuit?.listElements?.forEach { element ->
-                                    val obj = element.getPins().
-                                            find { it.getPoint() == startPoint }?.getElement()
-                                    if (obj != null) {
-                                        val deleteDialog: AlertDialog = this.let {
-                                            val builder = AlertDialog.Builder(context)
-                                            builder.apply {
-                                                setTitle("Delete Element-$obj")
-                                                setMessage("Are you sure you want to delete this element?")
-                                                setPositiveButton("Yes") { dialog, _ ->
-                                                    renderHelper?.removeObject(obj)
-                                                    obj.getPins().forEach { pin ->
-                                                        pin.removeFromNet()
-                                                        circuit?.listPins?.remove(pin)
-                                                    }
-                                                    circuit?.listElements?.remove(obj)
-                                                    invalidate()
-                                                    dialog.dismiss()
-                                                }
-                                                setNegativeButton("No") { dialog, _ ->
-                                                    dialog.dismiss()
-                                                }
-
-                                            }
-                                            builder.create()
-                                        }
-                                        deleteDialog.show()
-                                        performClick()
-                                        return true
-                                    }
-                                }
-                            }
-                        }
-                        EditEvent.DELETE_NET -> {
-                            startPoint = Point((x / renderHelper?.step!!).toInt(),
-                                    (y / renderHelper?.step!!).toInt())
-                            if (renderHelper?.isThereNet(startPoint)!!) {
-                                val obj = circuit?.listNets?.find {
-                                    it.getPoint() == startPoint}!!
-
-                                val deleteDialog: AlertDialog = this.let {
-                                    val builder = AlertDialog.Builder(context)
-                                    builder.apply {
-                                        setTitle("Delete Net-$obj")
-                                        setMessage("Are you sure you want to delete this net?")
-                                        setPositiveButton("Yes") { dialog, _ ->
-                                            renderHelper?.removeObject(obj)
-                                            obj.getPins().forEach { pin ->
-                                                circuit?.listPins?.remove(pin)
-                                            }
-                                            circuit?.listNets?.remove(obj)
-                                            invalidate()
-                                            dialog.dismiss()
-                                        }
-                                        setNegativeButton("No") { dialog, _ ->
-                                            dialog.dismiss()
-                                        }
-
-                                    }
-                                    builder.create()
-                                }
-                                deleteDialog.show()
-                            }
-                        }
-                        EditEvent.DELETE_CONNECTION -> TODO()
-                    }
-
-                }
-            }
-            performClick()
-            return true
-        }
-
-        override fun performClick(): Boolean {
-            super.performClick()
-            return true
-        }
-
-        fun setCircuit(circuit: Circuit) {
-            this.circuit = circuit
-            invalidate()
-        }
-
-        fun changeEditEvent(editEvent: EditEvent){
-            this.editEvent = editEvent
-            drawTouch = false
-        }
-
-        enum class EditEvent{
-            VIEW,
-            ADD_ELEMENT,
-            ADD_NET,
-            ADD_CONNECTION,
-            MOVE_ELEMENT,
-            MOVE_NET,
-            DELETE_ELEMENT,
-            DELETE_NET,
-            DELETE_CONNECTION
-        }
     }
 
 }
