+++ app/src/main/res/values/strings.xml
@@ -4,6 +4,15 @@
     <string name="processing_data">Processing data</string>
     <string name="please_wait">Please wait....</string>
 
+    <string name="error_move">Can\'t place %1$s here!</string>
+    <string name="delete_element">Delete Element-%1$s</string>
+    <string name="sure_delete_element">Are you sure you want to delete this element?</string>
+    <string name="delete_net">Delete Net-%1$s</string>
+    <string name="sure_delete_net">Are you sure you want to delete this net?</string>
+
+    <string name="yes">Yes</string>
+    <string name="no">No</string>
+
     <!-- TODO: Remove or change this placeholder text -->
     <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
